first commit
this is game based the bullet hell mechanics of undertale and just shapes and beats.

I have succesfully made a character that follows your mouse for movement.

I have made a projectile that can spawn on a random point on the xaxis that bounces when it connects with the edge of the screen boundary.

I made it so every 5 seconds, there are more projectiles added, making a irregular pattern that is harder and funner to avoid.